+++ BUI/src/com/jmex/bui/text/AWTTextFactory.java
@@ -312,14 +312,14 @@ public class AWTTextFactory extends BTextFactory {
             public void render(Renderer renderer, int x, int y, float alpha) {
                 //bimage.render(renderer, x, y, alpha);
                 bimage.render(renderer,
-                        0, 0, bimage.getWidth(), bimage.getHeight(),
+                        0, 0, (int) bimage.getWidth(), (int) bimage.getHeight(),
                         x, y, alpha);
             }
 
             public void render(Renderer renderer, int x, int y, int w, int h, float alpha) {
                 //bimage.render(renderer, x, y, w, h, alpha);
                 bimage.render(renderer,
-                        0, 0, bimage.getWidth(), bimage.getHeight(),
+                        0, 0, (int) bimage.getWidth(), (int) bimage.getHeight(),
                         x, y, w, h, alpha);
             }
